special casing for local-cluster

+++ operators/multiclusterobservability/controllers/placementrule/placementrule_controller.go
@@ -192,7 +192,7 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	if installMetricsWithoutAddon {
 		obsAddonList.Items = append(obsAddonList.Items, mcov1beta1.ObservabilityAddon{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:      localClusterName,
+				Name:      obsAddonName,
 				Namespace: config.GetDefaultNamespace(),
 			},
 		})
@@ -243,24 +243,39 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	latestClusters := []string{}
 	staleAddons := []string{}
 	for _, addon := range obsAddonList.Items {
-		latestClusters = append(latestClusters, addon.Namespace)
-		staleAddons = append(staleAddons, addon.Namespace)
+		log.Info("Coleen obsAddonList.Items", "namespace", addon.Namespace, "name", addon.Name)
+		if installMetricsWithoutAddon {
+			latestClusters = append(latestClusters, "local-cluster")
+			staleAddons = append(staleAddons, "local-cluster")
+
+		} else {
+			latestClusters = append(latestClusters, addon.Namespace)
+			staleAddons = append(staleAddons, addon.Namespace)
+		}
 	}
 	for _, work := range workList.Items {
 		if work.Name != work.Namespace+workNameSuffix {
-			reqLogger.Info("To delete invalid manifestwork", "name", work.Name, "namespace", work.Namespace)
+			if work.Namespace == config.GetDefaultNamespace() {
+				continue
+			}
+			reqLogger.Info("Coleen To delete invalid manifestwork", "name", work.Name, "namespace", work.Namespace)
 			err = deleteManifestWork(r.Client, work.Name, work.Namespace)
 			if err != nil {
 				return ctrl.Result{}, err
 			}
 		}
 		if !slices.Contains(latestClusters, work.Namespace) {
+			if work.Namespace == config.GetDefaultNamespace() {
+				staleAddons = commonutil.Remove(staleAddons, work.Namespace)
+				continue
+			}
 			reqLogger.Info("To delete manifestwork", "namespace", work.Namespace)
 			err = deleteManagedClusterRes(r.Client, work.Namespace)
 			if err != nil {
 				return ctrl.Result{}, err
 			}
 		} else {
+			log.Info("Coleen latestClusters staleaddon remove", "namespace", work.Namespace, "name", work.Name)
 			staleAddons = commonutil.Remove(staleAddons, work.Namespace)
 		}
 	}
@@ -269,7 +284,12 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	// but the managedclusteraddon for observability will not deleted by the cluster manager, so check against the
 	// managedclusteraddon list to remove the managedcluster resources after the managedcluster is detached.
 	for _, mcaddon := range managedclusteraddonList.Items {
+		log.Info("Coleen managedclusteraddonList.Items", "namespace", mcaddon.Namespace, "name", mcaddon.Name)
 		if !slices.Contains(latestClusters, mcaddon.Namespace) {
+			if mcaddon.Namespace == config.GetDefaultNamespace() {
+				staleAddons = commonutil.Remove(staleAddons, mcaddon.Namespace)
+				continue
+			}
 			reqLogger.Info("To delete managedcluster resources", "namespace", mcaddon.Namespace)
 			err = deleteManagedClusterRes(r.Client, mcaddon.Namespace)
 			if err != nil {
@@ -362,7 +382,11 @@ func createAllRelatedRes(
 	currentClusters := []string{}
 	for _, ep := range obsAddonList.Items {
 		log.Info("Coleen obsAddonList.Items", "namespace", ep.Namespace, "name", ep.Name)
-		currentClusters = append(currentClusters, ep.Namespace)
+		if ep.Namespace == config.GetDefaultNamespace() {
+			currentClusters = append(currentClusters, localClusterName)
+		} else {
+			currentClusters = append(currentClusters, ep.Namespace)
+		}
 	}
 
 	// need to reload the template and update the the corresponding resources
@@ -429,14 +453,13 @@ func createAllRelatedRes(
 
 	failedDeleteOba := false
 	for _, cluster := range currentClusters {
-		log.Info("To delete observabilityAddon", "namespace", cluster)
-		if cluster != config.GetDefaultNamespace() {
-			err = deleteObsAddon(c, cluster)
-			if err != nil {
-				failedDeleteOba = true
-				log.Error(err, "Failed to delete observabilityaddon", "namespace", cluster)
-			}
+		log.Info("Coleen To delete observabilityAddon if cluster != defaultnamespace", "namespace", cluster)
+		err = deleteObsAddon(c, cluster)
+		if err != nil {
+			failedDeleteOba = true
+			log.Error(err, "Failed to delete observabilityaddon", "namespace", cluster)
 		}
+
 	}
 
 	if failedCreateManagedClusterRes || failedDeleteOba {
